requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
frozenlist==1.3.3
idna==3.4
kombu==5.2.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
prompt-toolkit==3.0.38
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
twilio==8.2.0
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
